Validate emails for fact check more robustly
It's not enough to split on `,` and chek each for an `@` as if the value
is something like `foo@bar.com baz@qux.com` it'll pass that validation
but not actually be a valid email.  There's no notification later that the
fact check hasn't really gone out so we must be more robust in checking
them before we claim success.

We also add a hint to the UI to suggest that multiple email addresses can
be used as long as they are comma separated.