Fix Service Sign In Presenter bug
The `service_sign_in` format considers `choose_sign_in` `description` to be an
optional field.  We found a bug in the Service Sign In Presenter where
`choose_sign_in` `description` is always included in the payload regardless of
whether it is provided in a service_sign_in YAML file, causing the
`service_sign_in` schema validation to fail when `description` is not provided.
This makes the inclusion of `choose_sign_in` `description` in the payload
dependent on its presence in the YAML file.