Add task to unpublish service_sign_in content as `gone`
We add a rake task that invokes the `ServiceSignInUnpublishService` and
unpublishes a `service_sign_in` content item with type `gone`.

Unlike the other rake tasks for `service_sign_in` we have decided to pass
`content_id` and `locale` as arguments instead of the YAML file as it seems odd
to use the raw content itself to figure out what the content item to unpublish
is. It should be straightforward for a developer to find the content item's
`content_id` and `locale` and pass these as arguments instead.