Add missing default value for start button text
Ensure that all unarchived editions in the DB have a value for the start
button text, which is a required field.

This field already had a default value (see
https://github.com/alphagov/publisher/pull/638) which means that the
correct button text was displayed even for older editions which were
created before the field was added. But because the default value was
missing in the DB, it caused validation errors when unpublishing an old
edition: https://govuk.zendesk.com/agent/tickets/2544691