Updated option and hint text
I've changed the text on the options from "Use XXXX" to "Sign in with XXXX". I've also removed the introductory sentence ("If you’re filing online for the first time...") and added a similar sentence as hint text under the 3rd option. 

These changes were made in response to some user feedback we'd received over the last couple of weeks.