Make the link check box work in smaller browser windows
For browser windows below 992px, the link check box would cause the
form fields to get squashed into the left hand side of the well and
become unusable. For breakpoints below 992px (`md` in Bootstrap terms),
disable the float and force the box to appear above the form.

This is possibly not the best design decision, but it’s causing issues
at present for one of our partially sighted users.