Update ReverseMarkdown to use convert
It appears the `ReverseMarkdown` method `parse` is no longer supported, and
instead `convert` is used. The upside of this is we no longer need complex
regex to insert line breaks. We also make use of Ruby's `strip` method to get
rid of any excess whitespace.

It also appears that `ReverseMarkdown` now converts `<em>` tags to underscores
instead of asterisks to denote emphasis, and as this still has the effect we
require we update a test to reflect this.