Add ability to modify ordered related items
We introduce a new field into the Edition tagging form with the ability
to modify existing related items, and add up to 5 base paths at a time
(non-JS variant).

To use the base path as the primary data set (rather than content IDs),
which would be an unfeasible solution, we use the expanded links
endpoint to retrieve more data for a given edition. This allows us to
display and modify base paths in the front end for this field, and use
the lookup_by_base_path endpoint to translate them back to content IDs
as the patch_links endpoint expects. All other fields remain unchanged.