Remove unused admin_url method
This seems to have been imported in
e4fcdbb5c900f6024510041b6970b8a26d6391ba along with the rest of the
govuk_content_models, but doesn't look to be used in Publisher.

I'm looking at this as part of the Plek related external/internal
routing changes, and this method uses Plek.