Add and remove related items base paths dynamically
This completes matching the functionality from Content Tagger to allow
users to add and remove related items dynamically, including a URL check
via a publishing API proxy (to request lookup_by_base_path).

Extensive jQuery is used to minimise modification of the form field,
allowing for both JS and non-JS implementations to work. The fields are
rebuilt with additional actions (such as drag and remove).