Use generative expanded links to return link version
Changes to taggings will check if a prior change has been made to
prevent any conflicts from occuring. However, since changing the
requests from /v2/links/:content_id to /v2/expanded-links/:content_id,
the response did not include a version and was defaulting to 0 - and
therefore always displaying an error such as 'Somebody changed the tags
before you could. Your changes have not been saved'.

The request type was changed to include base_paths in the response which
is rendered in the ordered_related_items fields, rather than displaying
unusable content_ids.

We need to generate the expanded-links at runtime to include versioning
to match the functionality in Content Tagger. More details can be found
in this commit [1] in publishing-api.

[1] https://github.com/alphagov/publishing-api/commit/dd68de2c8521984429dee3a8b9c0566363777813