Request the expanded links with the locale
The link expansion in the Publishing API is locale specific. To
accurately reflect the related links that will be shown on a page,
it's necessary to use the locale parameter when calling the
get_expanded_links endpoint.

Without this, for non-english content, links will be missing to
content only available in that non-english locale.