Fix favicon not showing
I previously broke this by responding a bit too aggressively to a
warning emitted from many of the tests:

```
DEPRECATION WARNING: The asset "favicon.png" is not present in the asset pipeline.Falling back to an asset that may be in the public folder.
This behavior is deprecated and will be removed.
To bypass the asset pipeline and preserve this behavior,
use the `skip_pipeline: true` option.
```

Turns out the asset was in the pipeline but it couldn't be found easily.
This can be resolved by amending the path to include
`govuk_admin_template/`.