Remove all instances of the unicode sequence \u2028
This unicode sequence is identified as a change due to recent
updates to the sanitize gem. The knock on effect of this is that
govspeak validation fails and the edition cannot be saved.
Removing the offending sequence should have no effect on the presentation
of the markdown as it's an artefact from particular text editors.