Process mail based on env var
The environment var RUN_FACT_CHECK_FETCHER is set per environment
to control whether we fetch fact check emails. The previous
implementation used this var to determine whether or not to set the
imap details but not to prevent attempts to process the emails on the
server. So Mail#all would revert to defaults as no imap details were set
and this would result in a connection refused error for localhost:110.
The commit alters the behaviour so that we don't attempt to call
Mail#all if the env var is not set.