Add option to UpdateWorker to also publish
This allows the UpdateWorker to run an update and then a publish as a
synchronous pair of actions.

This is intended to replace the process where UpdateWorker and
PublishWorker are both ran asynchronously as this leads to out of order
issues where the UpdateWorker runs the UpdateService after the
PublishWorker has run PublishService. This out of ordering leads to an
out of document being published and a draft being left in Publishing
API.