Create experimental "knowledge API"
Running this rake task creates a new content item at
https://www.gov.uk/api/content/knowledge-alpha. It is currently
populated with the content of a number of high-profile guides in a
structured format.

- Minimum wage rates from https://www.gov.uk/national-minimum-wage-rates
- Income Tax rates and bands from https://www.gov.uk/income-tax-rates
- Self Assessment tax return deadlines from
https://www.gov.uk/self-assessment-tax-returns/deadlines

This is part of an experiment. We're currently working with one of the
voice providers (starts with an A) to see if we can get GOV.UK facts
directly into the knowledge engine that underpins the service. We've
checked this data format with them, and they said that they would be
able to consume this.

Obviously this isn't a very scalable solution - the information from
the guides is currently duplicated. Once the data changes we'll have to
update these numbers as well (a process for this will be set up).

A better idea might be to either 1) extract the data from the content,
2) manage the data elsewhere and inject it into the content. This would
require a lot of initial development however, so this initial "API"
serves as a test case. If we find that the voice provider won't be able
to use it, we can remove this content item easily.

This commit requires
https://github.com/alphagov/govuk-content-schemas/pull/772 to be
deployed.

https://trello.com/c/4mRr4Mzi