Include archived editions in major updates query
To determine the `public_updated_at` of an edition, we find the
latest major update of the artefact and use the `updated_at` of
that date. Unfortunately, the query is missing editions in the
archived state, meaning that they were once published but have
since been superseeded. We need to include these in the query to
make sure that we find the most recently `updated_at`, at the
moment it will miss any archived editions (which will be all the
previous editions) and always get the `updated_at` from the first
published edition.