Configure govuk_test and refactor helper methods
* Removes a bunch of test helpers which are either unused or
are tested by the jasmine specs.
* Removes '#trigger' on save button as this is not supported by
selenium and the 'element overlapping' issue no longer seems to be
a problem.
* Removes ambiguous 'with_ajax?' boolean which was modified to alias
'using_javascript?', possibly unintentionally.

The Jasmine test suite covers the scenarios for saving an edition
and the save button becoming disabled. Both of these scenarios have
timing related challenges in Capybara and led to flaky tests.