Add workaround for saving editions when save button is obscured
The workflow message element appears briefly after attempting
to save an edition. This may relate that the edition was saved
or that there was a problem saving.
There's a timing issue with clicking on the save button in
Capybara as the workflow message may be obscuring the save button
in which case we trigger a click using javascript.