Add basic question and answer parsing
This is part of the #govuk-voice project.  Content designers can mark up
existing content using the {answer} tags.  There should be a corresponding
{question} which allows them to be paired up.

This parser is used to both extract the question and answer pairs from the
content, and remove that markdown from the edition body so that it is not
sent to the publishing-api.

The extracted question and answer pairs will be extracted from each published
edition and combined into a list of question and answers which will be exposed
via the [knowledge api](https://github.com/alphagov/publisher/blob/master/app/services/knowledge_api.yml)