Allow variants of transaction pages
This commit adds a new concept named "variants" to transaction pages. Variants are technically similar to "parts", which are used for guides and other document types.

For transaction pages, variants allow transaction pages to respond to URLs of the form www.gov.uk/transaction/variant and display different page text and button targets based on the slug.

The primary use of this will initially be to allow GOV.UK Verify journeys which lead directly to a single IDP.