HUB-378: Update Verify redirect links
The Verify Hub supports journey hints to redirect users to the right journey flows
(registration, sign-in or eIDAS). However, the required hints are not being
implemented/passed by the services. We have developed a mechanism which
allows us to record the hint value (the option they selected on the GOV.UK
sign-in pages) by 'bouncing' the user via Verify first.