Use the correct capybara assertions
Refuting a capybara matcher is quite slow.  When we use `page.has_css?` on a
page that we don't expect it (so we're using `refute` as the assertion), capybara
waits for a while to see if it gets added at some point later.

If instead, we `assert` on the css _not_ being there, then capybara can finish
its check much more quickly.