Solve linting issues
In some cases, some of the rubocop cops have been disabled, rather than
addressing the issues.

This is not the right time to address these issues more extensively. We
had a conversation about this in the Platform Heath team and we agreed
that refactoring this code is not a priority right now, but there is a
Trello card to capture this:
https://trello.com/c/VNiRPVTP/1385-solve-publisher-linting-issues