Delete drafts from pub-api on deletion
Previously Publisher only deleted a draft from publishing-api when it was
the first edition of an Artefact and didn't have any previous editions
as this callback fired when deleting the parent Artefact.

The commit changes the code so whenever a draft is deleted then the
draft is also deleted from publishing-api, this is the correct behaviour
as previously it meant old drafts would linger in publishing-api when
they've been removed from Publisher leading to strange errors when the
new draft and deleted draft didn't match in pub-api. An error seen
recently was due to mismatched URL's between old (deleted)
and new draft. I assume this just works and overwrites the deleted draft
when the url isn't altered for a new draft.

This code also allows the deletion of other publisher specific states,
like "ammends_needed", "ready", "in_review" etc but these appear as
drafts in pub-api and a user can delete these editions from the UI the
same as a plain draft so these should also be deleted from pub-api.

Also added a test for the change, this felt like a slightly odd place as
you're testing specific logic in the model but publisher doesn't test web
requests in models so here it must live.

Trello:
https://trello.com/c/w0VMqd8R/822-deleting-drafts-with-slug-redirects-in-publisher-does-not-remove-slug-redirects