Remove deleted parts from the DOM after successful AJAX requests
Deleted parts are marked as destroyed and hidden in the DOM. This results in subsequent requests still including the deleted information and causing database errors. This commit ensures that, after a successful POST request deletes a part, it is also removed from the DOM and not included in any future requests.