Move top-level methods inside rake tasks
This fixes a linting issue about top-level methods in rake tasks by
moving the associated code into the rake task(s) that use it. Previously
some of the code was bloated due to printing lots of 'progress' strings.
In order to keep the rake tasks small, I've removed most of this, and
it can always be added back if it turns out to be 'vital'.