Use a matching group to parse an ID in the subject
This makes the code simpler and also means that the subject pattern can
include `.*?`, which is necessary because the subject is expected to
include the edition title, rather than being static.

Although the address-matching code includes a method to generate an
address from an ID, this doesn't seem to be necessary for the subject
(or at least doesn't seem to be the right place/way to do it).