Make the ID prefix in the subject configurable
This is expected to be configured by Puppet.

The current code uses an email address pattern in which the environment
(production/staging/integration) is embedded, and uses this both to
parse incoming emails and generate outgoing emails (the environment is
part of an opaque string as far as the app is concerned).

However, here I've made the prefix configurable but not the subject as a
whole: partly because the subject was already defined in code and not
configurable through Puppet, and partly because trying to both parse and
generate from the same string is needlessly complex.

The prefix can be left blank (e.g., in production).