Copy in_beta field when cloning editions
A pain point we experienced with publisher was that the in_beta flag is
not persisted onto new editions. This meant that it was easy to lose the
beta status.

I'm assuming that the omission of this field just represents a simple
oversight rather than any concious decision to exclude this.