Use mail-notify gem to deliver via GOV.UK Notify
This is just reapplying the changes that were made previously(cf. #1238, #1249),
taking into account what we've subsequently learned about the fact-check
workflow and Notify's reply-to behaviour. In particular: this no longer
sets the from or reply-to addresses, which must be set in the Notify
preferences instead.

This also involved some alterations to the template evaluation and
tests. We had to start explicitly making the helpers available to the
template --- it's not clear why this had succeeded previously. It also
means stubbing the bank holidays API endpoint, as this gets requested
when evaluating the template.