Improve regex for parsing an edition ID from a subject header
Previously the regex would look for an exact match based on what was
sent, pinned to the end of the string but not the start (to allow for
'Re:' and similar prefixes). However, it seems that in some cases there
can also be text added to the end of the string (e.g., an ID added by
the other party's own ticketing system) or even a complete modification
of the subject, in which case all we can do is ask the user not to
remove the ID.

I've made a few guesses as to what is an appropriate level of
cautiousness: for example, this looks for any hexadecimal string
surrounded by square brackes (and prefixed with an environment, if one
is set); however, I've considered it to be an error if there's more than
one match (because we can't be sure we'll pick the right one).

We could also try multiple ways of matching (e.g., look for a more exact
match with the subject we sent, and if it fails then any hex string; or
look for any hex string and if there are multiple matches then pick the
one that seems to be immediately prefixed by our subject, if any). But
since we don't know enough about the ways this could fail, it seemed
like a premature optimisation; it would be better to revisit this if we
start to see a common kind of failure in the future.