Read reply-to address from a separate environment variable
This allows us to stop showing the ID in the reply-to address we're
giving to users, without actually yet disabling it for incoming
messages.

The caveat is, though, that it should be compatible with both the
`address_format` (at least, compatible with the existing gmail filters
that assume that address format) and matching the address that the
`reply_to_id` refers to. Neither of these is guaranteed to break things
if they're different, but inconsistency is likely to complicate things
and it's difficult or impossible to reliably ensure that they're
compatible.