Add some documentation on the fact check process
We've been working in this area recently (and even had an incident) and
having this documentation at the beginning would have been very helpful.

It's doesn't go into too much detail in terms of individual lines of
code as I wanted to make sure this documentation remains accurate even
if the code itself is refactored or changed.