Scan for only the edition ID in fact check response subject
In 568c72d8c98641977285fc6aa7b7909cfaaa1795, the regex for subject
matching was changed. This has an unintended effect of blocking fact
check responses being processed where the subject contained any other
numbers in square brackets.

Therefore requiring the pattern to be exactly 24 characters, in order to
improve matching of the string we are interested in. MongoDB gives all
records a 12-digit hexadecimal number (i.e. 24 characters when written
out) so also updated all tests to reflect this.