Add auth_bypass_id field
As part of the work to improve how auth bypass tokens are used, we want
to change the behaviour so we that the auth bypass token is separate to
the ID of the edition. This would allow us to change the auth bypass
token if we needed to invalidate all the existing tokens (for example if
the token was leaked).

I've added a Rake task to populate the field with the current value, but
going forwards from here the token will be a UUID. The Rake task takes
about 10 minutes to run.