Rename 'NoisyWorkflow' mailer to mean something
Previously (in 2011) we added a NoisyWorkflow class to send an email
about general actions in the app [1]. Since that time, more specific
actions have been added to the mailer. In general, it's name has no
relationship with the code, disguises the fact that it's mailer, and
suggests it does something to mitigate noise, which it doesn't. This
renames most occurrences of 'noise' to refer to the event notifications
this code was written to send.

[1]: https://github.com/alphagov/publisher/commit/163b0dea17ecf66dfa2e15bc3349b0992b19f39f