Rename MultiNoisyWorkflow into a proto-service
This removes the last references to 'noisy workflow', by recasting
the 'MultiNoisyWorkflow' class as a service. Ideally it would only
have a single method (as with many other services in this repo).
Moving the code is still a net improvement, since this class is not
a mailer, and could be refactored to act like a service.