Make test output more concise
Previously we printed a little essay for each test, which made it
really hard (scrolling) to search for failures amidst all the ones
that passed. This switches back to a simple "." for each pass,
which is consistent with our other apps.

Note: this also removes a "ci_reporter_minitest" gem, which appears
to be redundant - we don't use CI XML reports.