Remove support for email fact check reading edition ID from address
Publisher used to link an email fact check to an edition via an edition
ID provided in the email address.

This was replaced by instead preferring to match the edition ID from the subject
line, which was a requirement of moving to GOVUK Notify as our e-mail service

https://github.com/alphagov/publisher/pull/1256

Now that this preferred solution is in use, the old code can now be removed.