Rescue bad request error and display a message
Currently, when running in staging or integration, if a fact check or
other email type is requested to be sent to an address not on the allow
list, the user will receive a generic error; it's difficult to guess
without experience of Notify that this is the problem, especially as
most users would not know of the allow list or be able to view/edit it.
This will give some further indication.