Use govuk_editor in tests
This sets up the tests to have the required permissions in advance of
later commits where we add tests for when the user doesn't have the
right permissions.

I'd prefer to do this rather than make the default user factory come
with those permissions to make the tests explicit, but this does mean
most of the tests that interact with users needs to be updated.