Add login_as_welsh_editor
It's a convenience test method that authenticates as a user with the
`welsh_editor` permission. I also refactored the code to make it easier
to log in as any user.

This is part of the changes required to implement the new Welsh editors
signon permissions.

Trello card: https://trello.com/c/Hq2qgwiI/2155-8-give-permission-in-publisher-for-welsh-teams-in-departments-to-update-only-welsh-content