Require govuk_editor to create artefacts
We don't want welsh_editors being able to create new artefacts.
The helper method  ensures that the user is a `govuk_editor`
otherwise redirects to the root page or the `redirect_path`
passed as argument.

This is part of the changes required to implement the new Welsh editors
signon permissions.

Trello card: https://trello.com/c/Hq2qgwiI/2155-8-give-permission-in-publisher-for-welsh-teams-in-departments-to-update-only-welsh-content