Allow Welsh editors to create new Welsh editions
This allows Welsh editors to only create new editions if their artefact
is Welsh. This is achieved by adding a `before_action` callback to the
editions controller that calls the `has_editor_permissions?` method
in the User model.

The permissions check in `NewVersionProcessor` has been removed (along
with its test) as it's no longer useful. Checking permissions in a
`before_action` callback in the controller means the processor is
never even reached when the user hasn't the right permissions.

It also adds a `welsh` trait to the edition factory to be used in the
added tests.

This is part of the changes required to implement the new Welsh editors
signon permissions.

Trello card: https://trello.com/c/Hq2qgwiI/2155-8-give-permission-in-publisher-for-welsh-teams-in-departments-to-update-only-welsh-content

Co-authored-by: Alan Gabbianelli <Alan.Gabbianelli@digital.cabinet-office.gov.uk>
Co-authored-by: Edward Kerry <Edward.Kerry@digital.cabinet-office.gov.uk>