Allow Welsh editors to update Welsh editions
This allows Welsh editors to only update editions if their artefact
is Welsh. This is achieved by using the  `before_action` callback
in the editions controller.

This is part of the changes required to implement the new Welsh editors
signon permissions.

Trello card: https://trello.com/c/Hq2qgwiI/2155-8-give-permission-in-publisher-for-welsh-teams-in-departments-to-update-only-welsh-content

Co-authored-by: Alan Gabbianelli <Alan.Gabbianelli@digital.cabinet-office.gov.uk>
Co-authored-by: Edward Kerry <Edward.Kerry@digital.cabinet-office.gov.uk>