Stop logging fact check email errors to Sentry
We already send an Icinga alert when there is a problem processing
the email: https://github.com/alphagov/govuk-puppet/blob/ce4eb1c7acc732e920e68cc1c3acdc7e5f86f9fb/modules/govuk/manifests/apps/publisher/unprocessed_emails_count_check.pp

We don't need to log it in Sentry too. The question is, which is
the most appropriate vehicle for notifying developers about the
error?

I would argue that it is more important for the Icinga alert to
happen, as we have 2nd line developers actively allocated to
responding to such issues. Errors logged to Sentry (currently)
languish until a developer periodically checks Sentry. Given the
sometimes urgent nature of fact checking, it is better that we
tackle issues quickly when they arise.

However, it is still useful to log the error to the Rails logger,
so that we can search for it in Logit.