Prevent Welsh Editors from adding edition notes to history
This commit adds a before_action callback to the NotesController to
ensure Welsh editors are only able to add or resolve important notes
on Welsh editions.

It also adds a new spec file for the NotesController.

This is part of the changes required to implement the new Welsh editors
signon permissions.

Trello card: https://trello.com/c/Hq2qgwiI/2155-8-give-permission-in-publisher-for-welsh-teams-in-departments-to-update-only-welsh-content

Co-authored-by: Alan Gabbianelli <Alan.Gabbianelli@digital.cabinet-office.gov.uk>
Co-authored-by: Edward Kerry <Edward.Kerry@digital.cabinet-office.gov.uk>