Prevent Welsh Editors from scheduling publications
Since the controller action to schedule publications (`update`) already
has a `before_action` callback that checks permissions, the
functionality to limit Welsh editors from scheduling publications for
non-Welsh editions is already present.

This is part of the changes required to implement the new Welsh editors
signon permissions.

Trello card: https://trello.com/c/Hq2qgwiI/2155-8-give-permission-in-publisher-for-welsh-teams-in-departments-to-update-only-welsh-content

Co-authored-by: Alan Gabbianelli <Alan.Gabbianelli@digital.cabinet-office.gov.uk>
Co-authored-by: Edward Kerry <Edward.Kerry@digital.cabinet-office.gov.uk>