Prevent Welsh Editors from cancelling scheduled publications
This commit limits Welsh editors from cancelling scheduled publications
for editions unless the scheduled edition is Welsh. This is done by
using a `before_action` callback for the `progress` action.

Since this permission check is now managed at the controller level, the
code used previously in `CancelScheduledPublishingProcessor` has beed
deleted, along with its test.

This is part of the changes required to implement the new Welsh editors
signon permissions.

Trello card: https://trello.com/c/Hq2qgwiI/2155-8-give-permission-in-publisher-for-welsh-teams-in-departments-to-update-only-welsh-content