Prevent Welsh editors from requesting reviews
Since the controller action to request a review (`update`) already has a
`before_action` callback that checks permissions, the functionality to
limit Welsh editors from requesting a 2i review for non-Welsh editions
is already present.

This commit:
  - adds tests for this behaviour
  - removes unnecessary code in `RequestReviewProcessor` (and its test)
  - adds a new edition factory trait used in the tests
  - hides the button to request a 2i review for a non-Welsh edition

This is part of the changes required to implement the new Welsh editors
signon permissions.

Trello card: https://trello.com/c/Hq2qgwiI/2155-8-give-permission-in-publisher-for-welsh-teams-in-departments-to-update-only-welsh-content