Prevent Welsh editors from requesting amendments
Since the controller action to request amendments to a review (`update`)
already has a before_action callback that checks permissions, the
functionality to limit Welsh editors from requesting amendments to a 2i
review for non-Welsh editions is already present.

This commit only adds tests for it and removes unnecessary code in
`RequestAmendmentsProcessor`.

This is part of the changes required to implement the new Welsh editors
signon permissions.

Trello card: https://trello.com/c/Hq2qgwiI/2155-8-give-permission-in-publisher-for-welsh-teams-in-departments-to-update-only-welsh-content