Prevent Welsh editors from approving reviews
Since the controller action to approve a review (`update`) already has a
before_action callback that checks permissions, the functionality to
limit Welsh editors from approving a 2i review for non-Welsh editions is
already present.

This commit only adds tests for it and removes unnecessary code in
`ApproveReviewProcessor` (and its test).

This is part of the changes required to implement the new Welsh editors
signon permissions.

Trello card:
https://trello.com/c/Hq2qgwiI/2155-8-give-permission-in-publisher-for-welsh-teams-in-departments-to-update-only-welsh-content