Prevent Welsh Editors from publishing Editions
This commit tests that Welsh editors may only directly publish
an edition if the edition is Welsh.

Publishing is handled through the EditionController's #update
method, which has a necessary before_action callback already applied
to check for user permissions.

The previous check in `PublishProcessor` is now not used and has
therefore been deleted.

This is part of the changes required to implement the new Welsh editors
signon permissions.

Trello card: https://trello.com/c/Hq2qgwiI/2155-8-give-permission-in-publisher-for-welsh-teams-in-departments-to-update-only-welsh-content

Co-Authored-By: Alan Gabbianelli <Alan.Gabbianelli@digital.cabinet-office.gov.uk>
Co-Authored-By: Edward Kerry <Edward.Kerry@digital.cabinet-office.gov.uk>