Hide fact check buttons from Welsh editors
This commit hides various fact check progress buttons on
Editions that user does not have sufficient permissions for.

Where appropriate the information for the current state will
still be displayed, but with no action possible.

These buttons simply trigger a modal, which is responsible for making
a post request. This request is dealt with directly in the controller.

It also deletes an old test which is now redundant due to these changes
and the controller before_action filtering.