Remove need for REDIS_HOST and REDIS_PORT env vars
This is one of the few GOV.UK projects that uses this old approach to
configuring Redis - whereas most of the other ones use the approach
supported natively by the Redis gem [1] of using REDIS_URL. Removing
this will allow us to remove the need for REDIS_HOST and REDIS_PORT env
vars in our architecture and also allows this code to be simplified.

It's worth noting that the namespace attribute in the Redis config had
no effect (it's a Sidekiq argument for Redis not a native Redis one) and
that the app already has hardcoded Redis namespacing when it is used.

[1]: https://github.com/redis/redis-rb